Task duplication

See merge request markussp/idatt1002_2021_28!53

+++ src/main/java/ntnu/idatt1002/controllers/EditTaskController.java
@@ -24,7 +24,7 @@ import java.util.ArrayList;
  */
 public class EditTaskController {
 
-    private int id;
+    private long id;
     @FXML private TextField titleTextField;
     @FXML private TextArea descriptionTextArea;
     @FXML private MenuButton categoryMenu;
@@ -111,11 +111,7 @@ public class EditTaskController {
         categoryMenu.setText(UserStateService.getCurrentUserCategory());
     }
 
-    /**
-     * A method to set a Id
-     * @param id
-     */
-    public void setId(int id){
+    public void setId(long id){
         this.id = id;
     }
